requirements.txt
ib_insync>=0.9.86
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
yfinance>=0.2.28
pandas-market-calendars>=4.3.0
pyyaml>=6.0
python-dateutil>=2.8.2
pytz>=2023.3
sqlalchemy>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
reportlab>=4.0.0
tabulate>=0.9.0
schedule>=1.2.0
openpyxl>=3.1.0
